dom/animation/test/mozilla/test_event_listener_leaks.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1450271 - Test Animation event listener leak conditions</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/dom/events/test/event_leak_utils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script class="testbody" type="text/javascript">
// Manipulate Animation.  Its important here that we create a
// listener callback from the DOM objects back to the frame's global
// in order to exercise the leak condition.
async function useAnimation(contentWindow) {
  let div = contentWindow.document.createElement("div");
  contentWindow.document.body.appendChild(div);
  let animation = div.animate({}, 100 * 1000);
  is(animation.playState, "running", "animation should be running");
  animation.onfinish = _ => {
    contentWindow.finishCount += 1;
  };
}

async function runTest() {
  try {
    await checkForEventListenerLeaks("Animation", useAnimation);
  } catch (e) {
    ok(false, e);
  } finally {
    SimpleTest.finish();
  }
}

SimpleTest.waitForExplicitFinish();
addEventListener("load", runTest, { once: true });
</script>
</pre>
</body>
</html>
